Migrate RouterCerts e2e test to Ginkgo framework
Migrates the TestRouterCerts e2e test to use the Ginkgo testing framework
as part of the broader test migration effort. This enables the test to be
run via the openshift-tests-extension binary and integrated with the OTE
test infrastructure.

Changes:
- Add test/e2e/certs.go with Ginkgo test implementation tagged with [Operator][Certs][Serial]
- Update test/e2e/certs_test.go to call the migrated test
- Fix Kubernetes secret naming to comply with RFC 1123 subdomain requirements
- Modify test/library/certs.go to accept testing.TB interface for compatibility
- Modify test/library/client.go to accept testing.TB interface for compatibility
- Modify test/library/waits.go to accept testing.TB interface for compatibility
- Add github.com/onsi/ginkgo/v2 as a direct dependency in go.mod
- Register e2e tests in cmd/cluster-authentication-operator-tests-ext/main.go

The test maintains the same functionality as the original implementation
while conforming to the Ginkgo test structure with proper tagging and
OTE framework integration.

Test Results:
✓ TestRouterCerts - PASSED (7.324s) via OTE binary
✓ TestRouterCerts - PASSED (4.42s) via make test-e2e
